+++ lib/active_validators/active_model/validations/postal_code_validator.rb
@@ -10,7 +10,7 @@ module ActiveModel
             country = 'us'
           end
         end
-        @formats = PostalCodeValidator.known_formats[country.to_s]
+        @formats = PostalCodeValidator.known_formats[country.to_s.downcase]
         raise "No known postal code formats for country #{country}" unless @formats
         record.errors.add(attribute) if value.blank? || !matches_any?
       end